test endpoints

+++ internal/models/order.go
@@ -23,8 +23,8 @@ type Order struct {
 	Seats       string      `gorm:"type:text;not null" json:"seats"` // "1-1,2-2,3-3"
 	TotalAmount float64     `gorm:"type:decimal(10,2);not null" json:"total_amount"`
 	Status      OrderStatus `gorm:"type:varchar(20);not null;default:'pending'" json:"status"`
-	Coins       *uint64     `gorm:"type:decimal(10,2);nullable" json:"coins"`
-	CoinsToPlus *uint64     `gorm:"type:decimal(10,2);nullable" json:"coins_to_plus"`
+	Coins       *float64    `gorm:"type:decimal(10,2);nullable" json:"coins"`
+	CoinsToPlus *float64    `gorm:"type:decimal(10,2);nullable" json:"coins_to_plus"`
 
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
